+++ src/main/java/org/mjulikelion/engnews/controller/AuthController.java
@@ -4,8 +4,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.mjulikelion.engnews.authentication.JwtEncoder;
-import org.mjulikelion.engnews.dto.request.LoginDto;
-import org.mjulikelion.engnews.dto.request.SignupDto;
+import org.mjulikelion.engnews.dto.request.auth.LoginDto;
+import org.mjulikelion.engnews.dto.request.auth.SignupDto;
 import org.mjulikelion.engnews.dto.response.ResponseDto;
 import org.mjulikelion.engnews.dto.response.TokenResponseDto;
 import org.mjulikelion.engnews.service.AuthService;
