+++ src/main/java/org/mjulikelion/engnews/controller/WordLikeController.java
@@ -3,9 +3,9 @@ package org.mjulikelion.engnews.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.mjulikelion.engnews.authentication.AuthenticatedUser;
-import org.mjulikelion.engnews.dto.request.CreateWordLikeDto;
+import org.mjulikelion.engnews.dto.request.wordLike.CreateWordLikeDto;
 import org.mjulikelion.engnews.dto.response.ResponseDto;
-import org.mjulikelion.engnews.dto.response.WordLikeListResponseData;
+import org.mjulikelion.engnews.dto.response.wordLike.WordLikeListResponseData;
 import org.mjulikelion.engnews.entity.User;
 import org.mjulikelion.engnews.service.WordLikeService;
 import org.springframework.http.HttpStatus;
