+++ src/main/java/org/mjulikelion/engnews/service/AuthService.java
@@ -3,8 +3,8 @@ package org.mjulikelion.engnews.service;
 import lombok.RequiredArgsConstructor;
 import org.mjulikelion.engnews.authentication.AccessTokenProvider;
 import org.mjulikelion.engnews.authentication.PasswordHashEncryption;
-import org.mjulikelion.engnews.dto.request.LoginDto;
-import org.mjulikelion.engnews.dto.request.SignupDto;
+import org.mjulikelion.engnews.dto.request.auth.LoginDto;
+import org.mjulikelion.engnews.dto.request.auth.SignupDto;
 import org.mjulikelion.engnews.dto.response.TokenResponseDto;
 import org.mjulikelion.engnews.entity.User;
 import org.mjulikelion.engnews.repository.UserRepository;
